Clean up AppDelegate
